Added UI.php to project
For convenience I added the required UI.php file that is included when
you download Chatscript.  It is required in the same folder as your
index.php or avatar1.php file in order to establish a connection to the
Chatscript server.   Please see the chatscript server instructions and
comments inside the file for configuratoin and setup.